[Add] userProfile route and controller
