+++ server/app.js
@@ -32,6 +32,7 @@ app.use(cors({ origin: true, credentials: true }));
 import { router as chatRoute } from "./routes/chat_route.js";
 import { router as userAuthRoute } from "./routes/userAuth_route.js";
 import { router as communityRoute } from "./routes/community_route.js";
+import { router as userProfileRoute } from "./routes/userProfile_route.js";
 
 app.get("/check", (req, res) => {
   console.log("health checking...");
@@ -42,6 +43,7 @@ app.use("/api/" + process.env.API_VERSION, [
   chatRoute,
   userAuthRoute,
   communityRoute,
+  userProfileRoute,
 ]);
 
 // Error handling
