+++ server/database.js
@@ -1,3 +1,4 @@
+// export connection to app.js
 import mongoose from "mongoose";
 import dotenv from "dotenv";
 dotenv.config();
